augmentation_tool/requirements.txt
click==8.1.3
Pillow==9.2.0
requests==2.28.1